requirements.txt
python-dotenv
uagents
openai
flask
psycopg2-binary
sqlalchemy
pymongo
gunicorn